backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
pdfplumber
pydantic[email]